Add dated log entries
After discussing with a friend on logs in experimental aircraft (and
how poor the logs are for most experimental aircraft), I have decided
to do my small part improving that.

In addition to the more prose-y version of what was done, I also have
a list of what work was done, along with when that work happened.

In that vein, this also creates pages for chapter 5 (I started
building some of the jigs), chapter 10 and chapter 11 (I received the
cut foam for the canard and elevators).